Create inline parser test module
